+++ lib/screens/home_inspection_screen/components/room_detail_model_form.dart
@@ -66,7 +66,7 @@ class _RoomDetailModelFormState extends State<RoomDetailModelForm> {
                 : '1',
             validator: storeyNumberValidation,
             onSaved: (newValue) {
-              roomDetail.storeyNo = int.tryParse(newValue);
+              roomDetail.storeyNo = double.tryParse(newValue);
             },
           ),
           CubitBuilder<RoomPurposeCubit, RoomPurposeState>(
